add checkZeroArgConstructor() and isJdk() function
